fix(opencv): size the normalize output in size_t, not int (CodeQL #362)
opencv_normalize computed the CHW array length as
float_img.rows * float_img.cols * 3 in int and only then widened it to
size_t. Past ~716 megapixels that overflows; the loop that fills the array
still writes rows * cols * 3 values, so a wrapped length is either a huge
allocation request (crash) or, past ~1.43 gigapixels, an array too small
for the loop (out-of-bounds write on the VM heap). CWE-190 -> CWE-787.

Reachable through Image->Normalize() on a program-built image whose
byte array matches those dimensions (opencv_raw_read checks the size in
size_t, so it is not affected). Widen before multiplying, as the ONNX
preprocessing already does ((size_t)3 * rgb.rows * rgb.cols in
onnx/eq/common.h).

+++ core/lib/opencv/opencv.cpp
@@ -1315,8 +1315,10 @@ extern "C" {
       channels[1] = (channels[1] - (float)mean_g) / (float)std_g;
       channels[2] = (channels[2] - (float)mean_b) / (float)std_b;
 
-      // Flatten to CHW format
-      const size_t total = float_img.rows * float_img.cols * 3;
+      // Flatten to CHW format. rows and cols are int: multiplying them first
+      // overflows past ~716 megapixels, and the loop below would then write
+      // rows * cols * 3 values into an array sized from the wrapped result.
+      const size_t total = static_cast<size_t>(float_img.rows) * static_cast<size_t>(float_img.cols) * 3;
       size_t* float_array = APITools_MakeFloatArray(context, total);
       double* float_array_buffer = reinterpret_cast<double*>(float_array + 3);
